Updated to Java 11+ as required by the Nashorn script engine Upgraded JavaFX to version 17.0.2 Changed to OpenJDK Nashorn script engine, version 15.3 Updated slf4j to 1.7.32 version Changed to javax.mail version 1.4.7 Upgraded to Java Service Wrapper 3.5.43 for Windows and 3.5.48 for Linux and Mac OSX Updated Vaadin to version 8.14.3 Updated PostgreSQL to version 42.3.3 Upgraded Milo to version 0.6.4 Upgraded Jetty to version 9.4.45.v20220203 Upgraded Jackson to version 2.13.1 Upgraded RabbitMQ to version 5.14.2 Upgraded ActiveMQ to version 5.16.4 Upgraded Modbus client to version 3.1.1 Upgraded Kafka client to version 3.1.0

+++ src/main/java/org/point85/domain/schedule/Team.java
@@ -152,8 +152,8 @@ public class Team extends Named implements Comparable<Team> {
 			return Duration.ZERO;
 		}
 
-		float days = (float) getRotation().getDuration().toDays();
-		float secPerWeek = (float) getRotation().getWorkingTime().getSeconds() * (7.0f / days);
+		float days = getRotation().getDuration().toDays();
+		float secPerWeek = getRotation().getWorkingTime().getSeconds() * (7.0f / days);
 		return Duration.ofSeconds((long) secPerWeek);
 	}
